requirements.txt
Babel==2.7.0
boto3==1.4.6
botocore==1.6.8
cairocffi==1.1.0
CairoSVG==2.4.2
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
cssselect2==0.2.2
defusedxml==0.6.0
distro==1.4.0
dj-database-url==0.5.0
Django==2.2.10
django-crispy-forms==1.8.0
django-heroku==0.3.1
django-paypal==1.0.0
django-phonenumber-field==3.0.1
django-sslserver==0.22
django-storages==1.8
docutils==0.15.2
gunicorn==20.0.4
html5lib==1.0.1
idna==2.8
jmespath==0.9.4
phonenumbers==8.10.21
Pillow==7.0.0
pkg-resources==0.0.0
psycopg2==2.8.3
pycparser==2.19
Pyphen==0.9.5
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
s3transfer==0.1.13
six==1.13.0
sqlparse==0.3.1
tinycss2==1.0.2
urllib3==1.25.7
uuid==1.30
WeasyPrint==50
webencodings==0.5.1
whitenoise==5.0.1
